Avancement hors de chez moi

+++ src/main/java/wishlist/dto/GroupDTO.java
@@ -18,7 +18,7 @@ public class GroupDTO {
     private String description;
 
     @NotNull
-    private Integer created_by;
+    private Long created_by;
 
     @NotNull
     @Size(min = 1, max = 10)
